Move test Mock consumer to the right place
